requirements.txt
Flask>=2.0.1
Werkzeug>=2.0.1
cryptography>=35.0.0
Flask-SocketIO>=5.1.1
pyotp>=2.6.0
qrcode>=7.3.1
pillow>=8.3.2
python-engineio>=4.3.0
python-socketio>=5.5.0
eventlet>=0.33.0
gunicorn>=20.1.0
